add-error-airport-behavior

+++ TemporalAirlinesConcept.Services/Implementations/Purchase/PurchaseActivities.cs
@@ -33,9 +33,6 @@ namespace TemporalAirlinesConcept.Services.Implementations.Purchase
 
             var flight = await flightHandle.QueryAsync(wf => wf.GetFlightDetails());
 
-            if (string.Equals(flight.From, Airports.ErrorCode) || string.Equals(flight.To, Airports.ErrorCode))
-                throw new Exception("Artificial error exception");
-
             if (flight.Seats.Count - flight.Registered.Count < 1)
                 return false;
 
@@ -180,7 +177,12 @@ namespace TemporalAirlinesConcept.Services.Implementations.Purchase
         [Activity]
         public async Task<bool> SaveTicketCompensationAsync(Ticket ticket)
         {
-            await _ticketRepository.DeleteTicketAsync(ticket.Id);
+            var ticketToDelete = await _ticketRepository.GetTicketAsync(ticket.Id);
+
+            if (ticketToDelete != null)
+            {
+                await _ticketRepository.DeleteTicketAsync(ticket.Id);
+            }
 
             return true;
         }
@@ -194,6 +196,12 @@ namespace TemporalAirlinesConcept.Services.Implementations.Purchase
             return true;
         }
 
+        [Activity]
+        public async Task<bool> ConfirmWithdrawCompensation()
+        {
+            return true;
+        }
+
         /// <summary>
         /// Retrieves the last flight from the given list of flight IDs.
         /// </summary>
@@ -201,7 +209,12 @@ namespace TemporalAirlinesConcept.Services.Implementations.Purchase
         [Activity]
         public async Task<DAL.Entities.Flight> GetFlightAsync(string flightId)
         {
-            return await _flightRepository.GetFlightAsync(flightId);
+            var flight = await _flightRepository.GetFlightAsync(flightId);
+
+            if (string.Equals(flight.From, Airports.ErrorCode) || string.Equals(flight.To, Airports.ErrorCode))
+                throw new Exception("Artificial error exception");
+
+            return flight;
         }
     }
 }
